requirements.txt
# VDK installation
quickstart-vdk
vdk-trino

# Necessary for sql in jupyter
ipython-sql
jupyterlab_sql

# Other
streamlit
jupyter-server-proxy
nbserverproxy

# Necessary for data job
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.3.5
numpy==1.21.5
datefinder==0.7.1
scipy==1.7.3
matplotlib==3.5.0
